mqtt2job -- initial commit
This is a minimal POC that enables the running of scheduled (or
triggered by some other condition) job(s) on a server, docker
container, or anything running perl, via an MQTT message sent
to a particular topic.

It is not intended to be secure in any way, at least in it's
current form. There is some basic checking that the command
exists in an assigned directory, but other security mechanisms
have yet to be added. So subject to usual DOS attacks etc.

Of particular note is the usage of the Home Assistant system
to trigger and monitor jobs sent via MQTT to the listener.

YAML templates can be generated by adding a few additional
options to the script instantiation (in base dir) as follows:

perl -Ilib bin/mqtt2job --mqtt_server mqtt.example.com \
     --base_topic mqtt_test --job_dir ./t \
     --ha_helper --cmd 00-run.t --task run_test_script

Note:
Remove options from --ha_helper onwards to activate listener
mode. This will ultimately run the first test in the "test
suite".