Mild refactor, moving all template generation operations to the library.
Also allow a handy mechanism for passing in an  argument to
vary the length of time it takes for a test to run.

Renamed use_perl to shebang as I think this is more understand-
able. Also allowed the process name to be set via a flag.

Finally, added some more todo idea.